Turn utility_* into stand-alone files
