+++ src/common/common.cmake
@@ -17,4 +17,18 @@ set(common_SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/syncfilestatus.cpp
 )
 
+if(WIN32)
+    list(APPEND common_SOURCES
+        ${CMAKE_CURRENT_LIST_DIR}/utility_win.cpp
+    )
+elseif(APPLE)
+    list(APPEND common_SOURCES
+        ${CMAKE_CURRENT_LIST_DIR}/utility_mac.mm
+    )
+elseif(UNIX AND NOT APPLE)
+    list(APPEND common_SOURCES
+        ${CMAKE_CURRENT_LIST_DIR}/utility_unix.cpp
+    )
+endif()
+
 configure_file(${CMAKE_CURRENT_LIST_DIR}/vfspluginmetadata.json.in ${CMAKE_CURRENT_BINARY_DIR}/vfspluginmetadata.json)
